[Wave] Remove unused code from iree/turbine/support
Also move used functions from the deleted files
into where they are being used.
